requirements.txt
obs-websocket-py
python-dotenv
psutil
pygetwindow
sounddevice
numpy
lancedb
pywin32; platform_system == "Windows"
google-generativeai
openai
httpx
pydub
